+++ geopandas/tests/test_geodataframe.py
@@ -1321,7 +1321,9 @@ class TestConstructor:
 
         geoms = [Point(1, 1), Point(2, 2), Point(3, 3)]
         gs = GeoSeries(geoms, name="my_geom")
-        gdf = GeoDataFrame({"a": [1, 2, 3]}, geometry=gs)
+        match = "You have passed a Series named "
+        with pytest.warns(UserWarning, match=match):
+            gdf = GeoDataFrame({"a": [1, 2, 3]}, geometry=gs)
 
         check_geodataframe(gdf, geometry_column="geometry")